# Conflicts:
#	src/main/java/ru/skypro/homework/dto/Comment.java
#	src/main/java/ru/skypro/homework/dto/Login.java
#	src/main/java/ru/skypro/homework/entity/CommentEntity.java
#	src/main/java/ru/skypro/homework/mapper/CommentMapper.java
#	src/main/java/ru/skypro/homework/repository/CommentRepository.java
#	src/main/resources/application.properties
#	src/main/resources/liquibase/changelog-master.yaml
#	src/main/resources/liquibase/script/comment.sql

+++ src/main/java/ru/skypro/homework/repository/CommentRepository.java
@@ -1,8 +1,6 @@
 package ru.skypro.homework.repository;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.skypro.homework.dto.Comment;
 import ru.skypro.homework.entity.CommentEntity;
 
-public interface CommentRepository extends JpaRepository<CommentEntity, Integer> {
+public interface CommentRepository extends JpaRepository<CommentEntity,Long > {
 }
